+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **[New refactoring]** Merge If Statements
+
 ## [0.1.0] - 2019-07-23
 
 ### Fixed